@@ -1,71 +1,68 @@
 {
-   "name": "com.healthnumeracyproject.healthnumeracy",
-   "displayName": "HealthNumeracy",
-   "version": "1.0.1",
-   "license": "Apache-2.0",
-   "description": "Math Gizmo",
-   "main": "index.js",
+   "name": "ci",
+   "version": "1.0.0",
+   "license": "MIT",
    "scripts": {
       "ng": "ng",
-      "start": "ng serve  --open --base-href /gizmo/",
+      "start": "ng serve --base-href /gizmo/",
       "build": "ng build --output-path=last-build/ --configuration=production --base-href /gizmo/",
       "lint": "ng lint",
       "prod-build": "ng build --output-path=last-build/ --configuration=production --base-href /gizmo/",
-      "apk-build": "ng build --output-path=www/ --configuration=production --base-href /gizmo/",
       "test-build": "ng build --output-path=test-build/ --configuration=testing --base-href /gizmo/",
-      "test": "echo \"Error: no test specified\" && exit 1"
+      "cordova-build": "ng build --output-path=www/ --configuration=testing --base-href ./"
    },
-   "private": true,
-   "keywords": [
-      "ecosystem:cordova"
-   ],
-   "author": "Health Numeracy Team",
    "dependencies": {
-      "@angular/animations": "^8.2.14",
-      "@angular/cdk": "^8.2.3",
-      "@angular/common": "^8.2.14",
-      "@angular/compiler": "^8.2.14",
-      "@angular/core": "^8.2.14",
-      "@angular/flex-layout": "^8.0.0-beta.27",
-      "@angular/forms": "^8.2.14",
-      "@angular/material": "^8.2.3",
-      "@angular/platform-browser": "^8.2.14",
-      "@angular/platform-browser-dynamic": "^8.2.14",
-      "@angular/router": "^8.2.14",
-      "@types/jquery": "^3.3.32",
-      "@types/mathjax": "^0.0.36",
-      "ajv": "^6.11.0",
-      "angular-sortablejs": "^2.7.0",
-      "angular2-fontawesome": "^5.2.1",
-      "copy-webpack-plugin": "5.1.1",
-      "cordova-android": "^8.1.0",
-      "core-js": "~3.6.4",
+      "@angular/animations": "^10.0.11",
+      "@angular/cdk": "^10.1.3",
+      "@angular/common": "^10.0.11",
+      "@angular/compiler": "^10.0.11",
+      "@angular/core": "^10.0.11",
+      "@angular/flex-layout": "^10.0.0-beta.32",
+      "@angular/forms": "^10.0.11",
+      "@angular/material": "^10.1.3",
+      "@angular/platform-browser": "^10.0.11",
+      "@angular/platform-browser-dynamic": "^10.0.11",
+      "@angular/router": "^10.0.11",
+      "@fullcalendar/adaptive": "^5.3.0",
+      "@fullcalendar/angular": "^5.3.0",
+      "@fullcalendar/daygrid": "^5.3.0",
+      "@fullcalendar/interaction": "^5.3.0",
+      "@fullcalendar/scrollgrid": "^5.3.0",
+      "@fullcalendar/timegrid": "^5.3.0",
+      "ajv": "^6.12.4",
+      "chart.js": "^2.9.3",
+      "core-js": "~3.6.5",
       "cryptiles": "^4.1.3",
-      "expr-eval": "^1.2.3",
+      "expr-eval": "~1.2.3",
       "font-awesome": "^4.7.0",
       "hammerjs": "^2.0.8",
-      "jquery": "^3.4.1",
-      "mathjax": "2.7.7",
-      "ngx-device-detector": "^1.3.20",
-      "npm-check-updates": "^4.0.1",
-      "rxjs": "^6.5.4",
-      "sortablejs": "^1.10.2",
-      "systemjs": "^6.2.2",
+      "jquery": "^3.5.1",
+      "moment": "^2.27.0",
+      "ng-recaptcha": "^5.0.0",
+      "ng2-charts": "^2.3.3",
+      "ngx-device-detector": "^1.4.6",
+      "ngx-perfect-scrollbar": "^10.0.1",
+      "rxjs": "^6.6.2",
+      "systemjs": "^6.5.0",
+      "tslib": "^2.0.1",
       "url-parse": "^1.4.7",
-      "zone.js": "~0.9.1"
+      "zone.js": "~0.11.1"
    },
    "devDependencies": {
-      "@angular-devkit/build-angular": "^0.803.25",
-      "@angular/cli": "^8.3.25",
-      "@angular/compiler-cli": "^8.2.14",
-      "@angular/language-service": "^8.2.14",
-      "@types/node": "^13.7.0",
-      "codelyzer": "~5.2.1",
-      "cordova-plugin-whitelist": "^1.3.4",
-      "tar": "^5.0.5",
-      "ts-node": "^8.6.2",
-      "tslint": "~5.20.1",
-      "typescript": " >=3.1 < 3.5"
+      "@angular-devkit/build-angular": "~0.1000.7",
+      "@angular/cli": "^10.0.6",
+      "@angular/compiler-cli": "^10.0.11",
+      "@angular/language-service": "^10.0.11",
+      "@types/node": "^14.6.0",
+      "@types/jquery": "^3.5.1",
+      "@types/mathjax": "^0.0.36",
+      "copy-webpack-plugin": "^6.0.3",
+      "codelyzer": "~5.2.2",
+      "npm-check-updates": "^7.1.1",
+      "ts-node": "^8.10.2",
+      "tslint": "~6.1.3",
+      "tar": "^6.0.5",
+      "typescript": ">=3.9.2 and <4.0.0"
    },
    "optionalDependencies": {
       "fsevents": "^2.1.2",
