@@ -1,6 +1,6 @@
 ﻿import {Injectable} from '@angular/core';
 import {HttpClient, HttpHeaders} from '@angular/common/http';
-import {map, catchError, finalize} from 'rxjs/operators';
+import {map, catchError} from 'rxjs/operators';
 import {environment} from '../../environments/environment';
 import {Router} from '@angular/router';
 
@@ -18,20 +18,42 @@ export class HttpService {
         private authenticationService: AuthenticationService) {
     }
 
-    post(url: string, body: any, auth: boolean = true) {
+    get(url: string, auth: boolean = true, params = null) {
         if (auth) {
-            // add authorization header with jwt token
             this.headers = new HttpHeaders({
                 'Authorization': 'Bearer '
                     + this.authenticationService.token, 'Content-Type': 'application/json'
             });
         } else {
-            // add authorization header with jwt token
             this.headers = new HttpHeaders({'Content-Type': 'application/json'});
         }
+        return this.http.get(this.apiUrl + url, {headers: this.headers, params: params})
+            .pipe(
+                map((response: Response) => response['message']),
+                catchError((response: Response) => {
+                    if (response['status_code'] === 401 || response['error']['status_code'] === 401) {
+                        this.authenticationService.logout();
+                        this.router.navigate(['login']);
+                    }
+                    if (response['status_code'] === 453 || response['error']['status_code'] === 453) {
+                        localStorage.setItem('redirect_to', this.router.url + '');
+                        this.router.navigate(['profile']);
+                    }
+                    return response['message'];
+                })
+            );
+    }
 
-        // post to api
-        return this.http.post(this.apiUrl + url, body, {headers: this.headers})
+    post(url: string, body: any = null, auth: boolean = true, params = null) {
+        if (auth) {
+            this.headers = new HttpHeaders({
+                'Authorization': 'Bearer '
+                    + this.authenticationService.token, 'Content-Type': 'application/json'
+            });
+        } else {
+            this.headers = new HttpHeaders({'Content-Type': 'application/json'});
+        }
+        return this.http.post(this.apiUrl + url, body, {headers: this.headers, params: params})
             .pipe(
                 map((response: Response) => response['message']),
                 catchError((response: Response) => {
@@ -40,25 +62,20 @@ export class HttpService {
                         this.router.navigate(['login']);
                     }
                     return response['message'];
-                }),
-                finalize(() => {
                 })
             );
     }
 
-    get(url: string, auth: boolean = true) {
+    put(url: string, body: any = null, auth: boolean = true, params = null) {
         if (auth) {
-            // add authorization header with jwt token
             this.headers = new HttpHeaders({
                 'Authorization': 'Bearer '
                     + this.authenticationService.token, 'Content-Type': 'application/json'
             });
         } else {
-            // add authorization header with jwt token
             this.headers = new HttpHeaders({'Content-Type': 'application/json'});
         }
-        // get from api
-        return this.http.get(this.apiUrl + url, {headers: this.headers})
+        return this.http.put(this.apiUrl + url, body, {headers: this.headers, params: params})
             .pipe(
                 map((response: Response) => response['message']),
                 catchError((response: Response) => {
@@ -67,9 +84,30 @@ export class HttpService {
                         this.router.navigate(['login']);
                     }
                     return response['message'];
-                }),
-                finalize(() => {
                 })
             );
     }
+
+    delete(url: string, auth: boolean = true, params = null) {
+        if (auth) {
+            this.headers = new HttpHeaders({
+                'Authorization': 'Bearer '
+                    + this.authenticationService.token, 'Content-Type': 'application/json'
+            });
+        } else {
+            this.headers = new HttpHeaders({'Content-Type': 'application/json'});
+        }
+        return this.http.delete(this.apiUrl + url, {headers: this.headers, params: params})
+            .pipe(
+                map((response: Response) => response['message']),
+                catchError((response: Response) => {
+                    if (response['status_code'] === 401 || response['error']['status_code'] === 401) {
+                        this.authenticationService.logout();
+                        this.router.navigate(['login']);
+                    }
+                    return response['message'];
+                })
+            );
+    }
+
 }
