@@ -1,1 +1,3 @@
-﻿export * from './user';
+﻿export * from './class';
+export * from './role';
+export * from './user';
