@@ -1,6 +1,11 @@
 ﻿export class User {
+    user_id: number;
     username: string;
+    first_name: string;
+    last_name: string;
     email: string;
     password: string;
-    questionNum: number;
+    question_num: number;
+    role: string;
+    country_id: number;
 }
