@@ -1,7 +1,12 @@
 export * from './authentication.service';
-﻿export * from './http.service';
+export * from './http.service';
 export * from './topic.service';
 export * from './tracking.service';
+export * from './user.service';
 export * from './placement.service';
 export * from './http-listener.service';
 export * from './welcome.service';
+export * from './dashboard.service';
+export * from './classes-management.service';
+export * from './content.service';
+export * from './country.service';
