@@ -4,30 +4,51 @@ import { HttpService } from './http.service';
 
 @Injectable()
 export class TrackingService {
+
+    private appId = +localStorage.getItem('app_id');
+
     constructor(
         private http: HttpService) {
     }
 
     startLesson(lesson_id) {
-        // notify api about lesson start
-        if (lesson_id === -1) {
-          /** TODO: change this HARDCODED value to testoutstart! */
-          // return this.http.post('/testoutstart', '')
-          return this.http.post('/', '');
+        let url = '/lesson/' + lesson_id + '/start';
+        if (this.appId) {
+            url += '?app_id=' + this.appId;
+        }
+        if (lesson_id !== -1) {
+            return this.http.post(url, '');
         } else {
-            return this.http.post('/lesson/' + lesson_id + '/start', '');
+            return this.http.post('/', ''); // return this.http.post('/testoutstart', '')
         }
     }
 
     doneLesson(topic_id, lesson_id, start_datetime, weak_questions) {
-        // notify api about lesson done
         const request = { start_datetime: start_datetime,
             weak_questions: weak_questions };
         if (lesson_id === -1) {
-          return this.http.post('/topic/' + topic_id + '/testoutdone', request);
+            let url = '/topic/' + topic_id + '/testout/done';
+            if (this.appId) {
+                url += '?app_id=' + this.appId;
+            }
+            return this.http.post(url, request);
         } else {
-          return this.http.post('/lesson/' + lesson_id + '/done', request);
+            let url = '/lesson/' + lesson_id + '/done';
+            if (this.appId) {
+                url += '?app_id=' + this.appId;
+            }
+            return this.http.post(url, request);
         }
+    }
+
+    finishTestout(topic_id, lesson_id, start_datetime, weak_questions) {
+        const request = { lesson_id: lesson_id, start_datetime: start_datetime,
+            weak_questions: weak_questions };
+        let url = '/topic/' + topic_id + '/testout/done-lessons';
+        if (this.appId) {
+            url += '?app_id=' + this.appId;
+        }
+        return this.http.post(url, request);
     }
 
     getLastVisitedLesson(student_id) {
@@ -40,5 +61,13 @@ export class TrackingService {
 
     getLastVisitedUnit(student_id) {
         return this.http.get('/unit/last-visited/' + student_id);
+    }
+
+    trackQuestionAnswer(question_id, is_right_answer) {
+        let url = '/question/' + question_id + '/tracking';
+        if (this.appId) {
+            url += '?app_id=' + this.appId;
+        }
+        return this.http.post(url, {'is_right_answer': is_right_answer});
     }
 }
