@@ -24,8 +24,11 @@ export class UserService {
     public changeProfile(user: User) {
         const request = {
             name: user.username,
+            first_name: user.first_name,
+            last_name: user.last_name,
             email: user.email,
-            question_num: user.questionNum
+            question_num: user.question_num,
+            country_id: user.country_id
         };
         return this.http.post('/profile', request)
             .pipe(
@@ -42,6 +45,81 @@ export class UserService {
             confirm_password: confirmedPassword
         };
         return this.http.post('/profile', request)
+            .pipe(
+                catchError(error => {
+                    console.log(error);
+                    throw Error(error);
+                })
+            );
+    }
+
+    public changeApplication(appId: number) {
+        const request = {
+            app_id: appId,
+        };
+        return this.http.post('/profile/app/', request)
+            .pipe(
+                catchError(error => {
+                    console.log(error);
+                    throw Error(error);
+                })
+            );
+    }
+
+    public getToDos() {
+        return this.http.get('/profile/todo')
+            .pipe(
+                map((response: Response) => {
+                    return response['items'];
+                }),
+                catchError(error => {
+                    console.log(error);
+                    throw Error(error);
+                })
+            );
+    }
+
+    public getClasses() {
+        return this.http.get('/profile/classes')
+            .pipe(
+                catchError(error => {
+                    console.log(error);
+                    throw Error(error);
+                })
+            );
+    }
+
+    public getClassInvitations() {
+        return this.http.get('/profile/classes/invitations')
+            .pipe(
+                map((response: Response) => {
+                    return response['items'];
+                }),
+                catchError(error => {
+                    console.log(error);
+                    throw Error(error);
+                })
+            );
+    }
+
+    public subscribeClass(classId: number) {
+        const request = {
+            class_id: classId,
+        };
+        return this.http.post('/profile/classes/' + classId + '/subscribe', request)
+            .pipe(
+                catchError(error => {
+                    console.log(error);
+                    throw Error(error);
+                })
+            );
+    }
+
+    public unsubscribeClass(classId: number) {
+        const request = {
+            class_id: classId,
+        };
+        return this.http.post('/profile/classes/' + classId + '/unsubscribe', request)
             .pipe(
                 catchError(error => {
                     console.log(error);
