@@ -3,41 +3,78 @@ import {BrowserModule} from '@angular/platform-browser';
 import {FormsModule} from '@angular/forms';
 import {HttpClientModule, HTTP_INTERCEPTORS} from '@angular/common/http';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import {
-    MatInputModule, MatButtonModule, MatSelectModule,
-    MatIconModule, MatMenuModule, MatRadioModule,
-    MatDialogModule, MatProgressBarModule, MatSliderModule,
-    MatToolbarModule, MatCardModule, MatCheckboxModule, MatDividerModule
-} from '@angular/material';
-import {Angular2FontawesomeModule} from 'angular2-fontawesome/angular2-fontawesome';
 import {FlexLayoutModule} from '@angular/flex-layout';
-import {SortablejsModule} from 'angular-sortablejs';
 import {DeviceDetectorModule} from 'ngx-device-detector';
+import { PerfectScrollbarModule } from 'ngx-perfect-scrollbar';
+import { PERFECT_SCROLLBAR_CONFIG } from 'ngx-perfect-scrollbar';
+
+import {MatInputModule} from '@angular/material/input';
+import {MatButtonModule} from '@angular/material/button';
+import {MatSelectModule} from '@angular/material/select';
+import {MatIconModule} from '@angular/material/icon';
+import {MatMenuModule} from '@angular/material/menu';
+import {MatRadioModule} from '@angular/material/radio';
+import {MatDialogModule} from '@angular/material/dialog';
+import {MatProgressBarModule} from '@angular/material/progress-bar';
+import {MatSliderModule} from '@angular/material/slider';
+import {MatToolbarModule} from '@angular/material/toolbar';
+import {MatCardModule} from '@angular/material/card';
+import {MatCheckboxModule} from '@angular/material/checkbox';
+import {MatDividerModule} from '@angular/material/divider';
+import {MatButtonToggleModule} from '@angular/material/button-toggle';
+import {MatSortModule} from '@angular/material/sort';
+import {MatSnackBarModule} from '@angular/material/snack-bar';
+import {MatListModule} from '@angular/material/list';
+import {DragDropModule} from '@angular/cdk/drag-drop';
 
 import {AppRoutingModule} from './app.routing';
 import {AuthGuard} from './_guards/index';
 
-import {AuthenticationService, HttpService, HTTPListener, HTTPStatus} from './_services/index';
+import {AuthenticationService, HttpService, HTTPListener, HTTPStatus, CountryService} from './_services/index';
 
 import {AppComponent} from './_components/app.component';
-import {WelcomeComponent} from './_components/welcome/index';
-import {LoginComponent, ForgotPasswordComponent} from './_components/welcome/login/index';
-import {RegisterComponent} from './_components/welcome/register/index';
-import {TryComponent} from './_components/welcome/try/try.component';
-import {HomeComponent} from './_components/home/index';
-import {TopicComponent} from './_components/home/topic/index';
-import {
-    LessonComponent, GoodDialogComponent, BadDialogComponent,
-    ReportDialogComponent, FeedbackDialogComponent, BadChallengeDialogComponent, ChartComponent
-} from './_components/home/topic/lesson/index';
+import {WelcomeComponent, RegisterComponent, LoginComponent,
+    ForgotPasswordComponent, ResetPasswordComponent, TryComponent} from './_components/auth/index';
+import {AssignmentComponent} from './_components/assignment/index';
+import {TopicComponent} from './_components/assignment/topic/index';
+import {LessonComponent, ChartComponent} from './_components/assignment/topic/lesson/index';
 import {ProfileComponent} from './_components/profile/profile.component';
-import {ResetPasswordComponent} from './_components/welcome/login/reset-password/reset-password.component';
-import {QuestionComponent} from './_components/home/topic/lesson/question/question.component';
-
+import {GoodDialogComponent, BadDialogComponent, ReportDialogComponent,
+    FeedbackDialogComponent, BadChallengeDialogComponent, YesNoDialogComponent} from './_components/dialogs/index';
+import {ToDoComponent, MyClassesComponent, MyInvitationsComponent} from './_components/student/index';
+import {ClassReportComponent, ClassStudentsComponent,
+    StudentAssignmentsDialogComponent, AddStudentDialogComponent,
+    EditClassDialogComponent, ManageClassesComponent, ManageAssignmentsComponent,
+    EditAssignmentDialogComponent, ReviewContentComponent, ClassDashboardComponent,
+    ClassAssignmentsComponent, ClassAssignmentsCalendarComponent, ClassToDoComponent,
+    ClassMenuComponent, StudentsUsageChartComponent, ClassDetailedReportComponent,
+    EditClassAssignmentDialogComponent} from './_components/teacher/index';
+import {DashboardComponent} from './_components/dashboard/dashboard.component';
+import {QuestionComponent} from './_components/assignment/topic/lesson/question/question.component';
 import {QuestionPreviewComponent} from './_components/previews/question-preview/question-preview.component';
-import {PlacementComponent, QuestionNumDialogComponent} from './_components/welcome/placement/index';
+// import {PlacementComponent, QuestionNumDialogComponent} from './_components/welcome/placement/index';
+
+import { RecaptchaModule, RecaptchaFormsModule } from 'ng-recaptcha';
 
 import {DraggableDirective} from './_directives/draggable.directive';
+import { TableFilterPipe, SafeHtmlPipe } from './_pipes/index';
+
+import { ChartsModule } from 'ng2-charts';
+
+import { FullCalendarModule } from '@fullcalendar/angular';
+import dayGridPlugin from '@fullcalendar/daygrid';
+import timeGridPlugin from '@fullcalendar/timegrid';
+import interactionPlugin from '@fullcalendar/interaction';
+import adaptivePlugin from '@fullcalendar/adaptive';
+import scrollGridPlugin from '@fullcalendar/scrollgrid';
+
+FullCalendarModule.registerPlugins([
+    dayGridPlugin,
+    timeGridPlugin,
+    interactionPlugin,
+    adaptivePlugin,
+    scrollGridPlugin
+]);
 
 @NgModule({
     imports: [
@@ -45,8 +82,9 @@ import {DraggableDirective} from './_directives/draggable.directive';
         FormsModule,
         HttpClientModule,
         AppRoutingModule,
-        Angular2FontawesomeModule,
         BrowserAnimationsModule,
+        FullCalendarModule,
+        ChartsModule,
         MatInputModule,
         MatButtonModule,
         MatSelectModule,
@@ -60,8 +98,15 @@ import {DraggableDirective} from './_directives/draggable.directive';
         MatCardModule,
         MatCheckboxModule,
         MatDividerModule,
+        MatButtonToggleModule,
+        MatSortModule,
+        MatSnackBarModule,
+        MatListModule,
+        DragDropModule,
         FlexLayoutModule,
-        SortablejsModule.forRoot({animation: 150}),
+        PerfectScrollbarModule,
+        RecaptchaModule,
+        RecaptchaFormsModule,
         DeviceDetectorModule.forRoot()
     ],
     exports: [],
@@ -70,7 +115,7 @@ import {DraggableDirective} from './_directives/draggable.directive';
         WelcomeComponent,
         LoginComponent,
         RegisterComponent,
-        HomeComponent,
+        AssignmentComponent,
         TopicComponent,
         LessonComponent,
         GoodDialogComponent,
@@ -80,26 +125,57 @@ import {DraggableDirective} from './_directives/draggable.directive';
         BadChallengeDialogComponent,
         ChartComponent,
         ProfileComponent,
+        ToDoComponent,
+        MyClassesComponent,
+        MyInvitationsComponent,
+        ManageClassesComponent,
+        YesNoDialogComponent,
+        EditClassDialogComponent,
+        StudentAssignmentsDialogComponent,
+        AddStudentDialogComponent,
+        ManageAssignmentsComponent,
+        EditAssignmentDialogComponent,
+        EditClassAssignmentDialogComponent,
+        ReviewContentComponent,
+        ClassDashboardComponent,
+        ClassAssignmentsComponent,
+        ClassAssignmentsCalendarComponent,
+        ClassToDoComponent,
+        ClassMenuComponent,
+        StudentsUsageChartComponent,
+        ClassDetailedReportComponent,
         TryComponent,
         ForgotPasswordComponent,
         ResetPasswordComponent,
         QuestionComponent,
         QuestionPreviewComponent,
-        PlacementComponent,
-        QuestionNumDialogComponent,
-        DraggableDirective
+        DashboardComponent,
+        ClassReportComponent,
+        ClassStudentsComponent,
+        // PlacementComponent,
+        // QuestionNumDialogComponent,
+        DraggableDirective,
+        TableFilterPipe,
+        SafeHtmlPipe
     ],
     entryComponents: [
         GoodDialogComponent,
         BadDialogComponent,
         ReportDialogComponent,
         FeedbackDialogComponent,
-        QuestionNumDialogComponent,
-        BadChallengeDialogComponent
+        // QuestionNumDialogComponent,
+        BadChallengeDialogComponent,
+        YesNoDialogComponent,
+        EditClassDialogComponent,
+        StudentAssignmentsDialogComponent,
+        AddStudentDialogComponent,
+        EditAssignmentDialogComponent,
+        EditClassAssignmentDialogComponent
     ],
     providers: [
         AuthGuard,
         AuthenticationService,
+        CountryService,
         HttpService,
         HTTPListener,
         HTTPStatus,
@@ -107,6 +183,12 @@ import {DraggableDirective} from './_directives/draggable.directive';
             provide: HTTP_INTERCEPTORS,
             useClass: HTTPListener,
             multi: true
+        },
+        {
+            provide: PERFECT_SCROLLBAR_CONFIG,
+            useValue: {
+                suppressScrollX: true
+            }
         },
     ],
     bootstrap: [AppComponent]
