@@ -5,29 +5,32 @@ namespace App\Http\APIControllers;
 use App\PlacementQuestion;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
-use JWTAuth;
+use Tymon\JWTAuth\Facades\JWTAuth;
 use App\Unit;
 
 class PlacementController extends Controller
 {
 
-    /**
-     * return placement questions.
-     *
-     * @return array
-     */
+    private $user;
+
+    public function __construct()
+    {
+        try {
+            $this->user = JWTAuth::parseToken()->authenticate();
+            if (!$this->user) {
+                abort(401, 'Unauthorized!');
+            }
+        } catch (\Exception $e) {
+            abort(401, 'Unauthorized!');
+        }
+    }
+
     public function get()
     {
         $placement = PlacementQuestion::all();
         return $this->success($placement);
     }
 
-    /**
-     * return first topic id.
-     *
-     * @param $unit_id
-     * @return int
-     */
     public function getTopicId($unit_id) {
         if(!$unit_id || !is_numeric($unit_id)) {
             return $this->error('id must be integer');
@@ -41,38 +44,32 @@ class PlacementController extends Controller
         return $this->success($topic_id);
     }
 
-    /**
-     * Mark a half of unit as done.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function doneHalfUnit(Request $request) {
-        $unit_id = $request->unit_id;
-        $student = JWTAuth::parseToken()->authenticate();
+        $unit_id = $request['unit_id'];
+        $student = $this->user;
         $topics = DB::table('topic')
             ->join('unit', 'topic.unit_id', '=', 'unit.id')
             ->where('unit.id', $unit_id)
             ->select('topic.id')
-            ->get();
+            ->get()->all();
         $middleTopicIndex = round(count($topics)/2);
         for ($i = 0; $i < $middleTopicIndex-1; $i++) {
             $topicId = $topics[$i]->id;
             DB::table('progresses')->insert([
                 'student_id' => $student->id,
                 'entity_id' => $topicId,
-                'entity_type' => 1
+                'entity_type' => 'topic'
             ]);
             $lessons = DB::table('lesson')
                 ->join('topic', 'lesson.topic_id', '=', 'topic.id')
                 ->where('topic.id', '=', $topicId)
                 ->select('lesson.id')
-                ->get();
+                ->get()->all();
             foreach ($lessons as $lesson) {
                 DB::table('progresses')->insert([
                     'student_id' => $student->id,
                     'entity_id' => $lesson->id,
-                    'entity_type' => 0
+                    'entity_type' => 'lesson'
                 ]);
             }
         }
@@ -80,27 +77,21 @@ class PlacementController extends Controller
         return $this->success($topic_id);
     }
 
-    /**
-     * Mark unit as done.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function doneUnit(Request $request) {
-        $unit_id = $request->unit_id;
-        $student = JWTAuth::parseToken()->authenticate();
+        $unit_id = $request['unit_id'];
+        $student = $this->user;
         // done lessons
         $lessons = DB::table('lesson')
             ->join('topic', 'lesson.topic_id', '=', 'topic.id')
             ->join('unit', 'topic.unit_id', '=', 'unit.id')
             ->where('unit.id', $unit_id)
             ->select('lesson.id')
-            ->get();
+            ->get()->all();
         foreach ($lessons as $lesson) {
             DB::table('progresses')->insert([
                 'student_id' => $student->id,
                 'entity_id' => $lesson->id,
-                'entity_type' => 0
+                'entity_type' => 'lesson'
             ]);
         }
         // done topics
@@ -108,37 +99,37 @@ class PlacementController extends Controller
             ->join('unit', 'topic.unit_id', '=', 'unit.id')
             ->where('unit.id', $unit_id)
             ->select('topic.id')
-            ->get();
+            ->get()->all();
         foreach ($topics as $topic) {
             DB::table('progresses')->insert([
                 'student_id' => $student->id,
                 'entity_id' => $topic->id,
-                'entity_type' => 1
+                'entity_type' => 'topic'
             ]);
         }
         // done unit
         DB::table('progresses')->insert([
             'student_id' => $student->id,
             'entity_id' => $unit_id,
-            'entity_type' => 2
+            'entity_type' => 'unit'
         ]);
-        
+
         //find all units from level that are not done yet
         $unit_model = Unit::where("id", $unit_id)->first();
         $units = DB::table('unit')->leftJoin('progresses', function ($join) use ($student) {
             $join->on('progresses.student_id', '=', DB::raw($student->id))
-            ->on('progresses.entity_type', '=', DB::raw(2))
+            ->on('progresses.entity_type', '=', DB::raw('"unit"'))
             ->on('progresses.entity_id', '=', 'unit.id');
         })
         ->where(['level_id' => $unit_model->level_id, 'dependency' => 1])
-        ->whereNull('progresses.id')->get();
+        ->whereNull('progresses.id')->get()->all();
         //if all units are done, mark level as done
         if (!count($units)) {
             // done level
             DB::table('progresses')->insert([
                 'student_id' => $student->id,
                 'entity_id' => $unit_model->level_id,
-                'entity_type' => 3
+                'entity_type' => 'level'
             ]);
         }
 
