@@ -8,49 +8,39 @@ use App\Unit;
 
 class PlacementController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    public function __construct()
+    {
+        // $this->authorizeResource(PlacementQuestion::class); // not working!
+    }
+
     public function index(Request $request)
     {
-        if ($request->has('sort') and $request->has('order')) {
+        $this->checkAccess(auth()->user()->isSuperAdmin() || auth()->user()->isAdmin());
+        if ($request['sort'] && $request['order']) {
             $placements = PlacementQuestion::with('unit')
-                ->orderBy($request->sort, $request->order)->get();
+                ->orderBy($request['sort'], $request['order'])->get();
         } else {
             $placements = PlacementQuestion::with('unit')->get();
         }
-        return view('placement_views.index', ['placements'=>$placements]);
+        return view('placements.index', ['placements'=>$placements]);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
+        $this->checkAccess(auth()->user()->isSuperAdmin() || auth()->user()->isAdmin());
         $total_placements = PlacementQuestion::all()->count();
-        $placements = PlacementQuestion::with('unit')->get();
         $units = Unit::all();
         $lid = "";
-        return view('placement_views.create', array(
-            'placements' => $placements,
+        return view('placements.create', array(
             'total_placements' => $total_placements,
             'units' => $units,
             'lid' => $lid
         ));
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
+        $this->checkAccess(auth()->user()->isSuperAdmin() || auth()->user()->isAdmin());
         $this->validate($request, [
          'unit_id' => 'required'
          ]);
@@ -63,46 +53,30 @@ class PlacementController extends Controller
         $placement->save();
         $placements = PlacementQuestion::with('unit')->get();
         \Session::flash('flash_message', 'successfully saved.');
-        return view('placement_views.index', ['placements'=>$placements]);
+        return view('placements.index', ['placements'=>$placements]);
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function show()
     {
         return "Under Construction";
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit($id)
     {
+        $this->checkAccess(auth()->user()->isSuperAdmin() || auth()->user()->isAdmin());
         $placement = PlacementQuestion::find($id);
         $total_placements = PlacementQuestion::all()->count();
         $units = Unit::all();
-        return view('placement_views.edit', [
+        return view('placements.edit', [
             'placement' => $placement,
             'total_placements' => $total_placements,
             'units' => $units
         ]);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, $id)
     {
+        $this->checkAccess(auth()->user()->isSuperAdmin() || auth()->user()->isAdmin());
         $this->validate($request, [
          'unit_id' => 'required'
          ]);
@@ -113,20 +87,15 @@ class PlacementController extends Controller
         $unit = Unit::find($request['unit_id']);
         $placement->unit()->associate($unit);
         $placement->save();
-        return redirect('/placement_views')
+        return redirect('/placements')
             ->with(array('message'=> 'Updated successfully'));
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function destroy($id)
     {
+        $this->checkAccess(auth()->user()->isSuperAdmin() || auth()->user()->isAdmin());
         PlacementQuestion::where('id', $id)->delete();
-        return redirect('/placement_views')
+        return redirect('/placements')
             ->with(array('message'=> 'Deleted successfully'));
     }
 }
