@@ -4,58 +4,50 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Input;
 use App\Unit;
 use App\Level;
 
 class UnitController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @param Request $request
-     * @return \Illuminate\Http\Response
-     */
-    public function index(Request $request)
+    public function __construct()
     {
-        $levels = Level::all();
-
-        $query = Unit::query();
-
-        $query->when($request->has('level_id') && ($request->level_id >= 0), function ($q) {
-            return $q->where('level_id', request('level_id'));
-        });
-
-        $query->when($request->has('id'), function ($q) {
-            return $q->where('id', request('id'));
-        });
-        $query->when($request->has('order_no'), function ($q) {
-            return $q->where('order_no', request('order_no'));
-        });
-        $query->when($request->has('title'), function ($q) {
-            return $q->where('title', 'LIKE', '%'.request('title').'%');
-        });
-        $query->when($request->has('sort') and $request->has('order'), function ($q) {
-            return $q->orderBy(request('sort'), request('order'));
-        });
-
-        $units = $query->paginate(10)->appends(Input::except('page'));
-
-        return view('unit_views.index', ['levels'=>$levels, 'units'=>$units, 'level_id' => $request->level_id]);
+        // $this->authorizeResource(Unit::class); // not working!
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    public function index(Request $request)
+    {
+        $this->checkAccess(auth()->user()->isSuperAdmin() || auth()->user()->isAdmin());
+        $query = Unit::query();
+        if ($request['level_id']) {
+            $query->where('level_id', $request['level_id']);
+        }
+        if ($request['id']) {
+            $query->where('id', $request['id']);
+        }
+        if ($request['order_no']) {
+            $query->where('order_no', $request['order_no']);
+        }
+        if ($request['title']) {
+            $query->where('title', 'LIKE', '%' . $request['title'] . '%');
+        }
+        if ($request['sort'] && $request['order']) {
+            $query->orderBy($request['sort'], $request['order']);
+        }
+        return view('units.index', [
+            'levels' => Level::all(),
+            'units' => $query->paginate(10)->appends(request()->query()),
+            'level_id' => $request['level_id']
+        ]);
+    }
+
     public function create()
     {
+        $this->checkAccess(auth()->user()->isSuperAdmin() || auth()->user()->isAdmin());
         $lid = "";
         $levels = DB::select('select * from level');
         $units = DB::table('unit')->where('level_id', $lid)->get();
         $total_unit = Unit::all()->count();
-        return view('unit_views.create', [
+        return view('units.create', [
             'levels' => $levels,
             'units' => $units,
             'lid' => $lid,
@@ -63,78 +55,17 @@ class UnitController extends Controller
         ]);
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
+        $this->checkAccess(auth()->user()->isSuperAdmin() || auth()->user()->isAdmin());
         // Store topic title and unit_id into topic table
         $this->validate($request, [
-         'level_id'    => 'required',
-         'unit_title'=> 'required',
+            'level_id' => 'required',
+            'unit_title' => 'required',
         ]);
         DB::table('unit')->insert([
-         'title' => $request['unit_title'],
-         'dependency' => $request['dependency'] ?: false,
-         'dev_mode' => $request['dev_mode'] ?: false,
-         'level_id' => $request['level_id'],
-         'order_no' => $request['order_no'],
-         'created_at' => date('Y-m-d H:i:s'),
-         'modified_at' => date('Y-m-d H:i:s')
-        ]);
-        $level_id = $request->input('level_id');
-        return redirect('/unit_views?level_id='. $level_id)->with(array('message'=> 'Created successfully'));
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function show()
-    {
-        return "Under Construction";
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        $unit = DB::table('unit')
-            ->join('level', 'unit.level_id', '=', 'level.id')
-            ->select('unit.*', 'level.title as ltitle', 'level.id as lid')
-            ->where('unit.id', '=', $id)->first();
-        $levels = DB::table('level')->select('id', 'title')->where('id', $unit->lid)->get();
-        $total_unit = Unit::all()->count();
-        return view('unit_views.edit', [
-            'levels'=>$levels,
-            'unit'=>$unit,
-            'total_unit'=>$total_unit,
-        ]);
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, $id)
-    {
-        $this->validate($request, [
-            'level_id'    => 'required',
-            'unit_title'=> 'required'
-        ]);
-        DB::table('unit')->where('id', $id)->update([
             'title' => $request['unit_title'],
+            'description' => $request['description'],
             'dependency' => $request['dependency'] ?: false,
             'dev_mode' => $request['dev_mode'] ?: false,
             'level_id' => $request['level_id'],
@@ -143,20 +74,56 @@ class UnitController extends Controller
             'modified_at' => date('Y-m-d H:i:s')
         ]);
         $level_id = $request->input('level_id');
-        return redirect('/unit_views?level_id='. $level_id)->with(array('message'=> 'Updated successfully'));
+        return redirect('/units?level_id=' . $level_id)->with(array('message' => 'Created successfully'));
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param Request $request
-     * @param  int $id
-     * @return \Illuminate\Http\Response
-     */
+    public function show()
+    {
+        return "Under Construction";
+    }
+
+    public function edit($id)
+    {
+        $this->checkAccess(auth()->user()->isSuperAdmin() || auth()->user()->isAdmin());
+        $unit = DB::table('unit')
+            ->join('level', 'unit.level_id', '=', 'level.id')
+            ->select('unit.*', 'level.title as ltitle', 'level.id as lid')
+            ->where('unit.id', '=', $id)->first();
+        $levels = DB::select('select * from level');
+        $total_unit = Unit::all()->count();
+        return view('units.edit', [
+            'levels' => $levels,
+            'unit' => $unit,
+            'total_unit' => $total_unit,
+        ]);
+    }
+
+    public function update(Request $request, $id)
+    {
+        $this->checkAccess(auth()->user()->isSuperAdmin() || auth()->user()->isAdmin());
+        $this->validate($request, [
+            'level_id' => 'required',
+            'unit_title' => 'required'
+        ]);
+        DB::table('unit')->where('id', $id)->update([
+            'title' => $request['unit_title'],
+            'description' => $request['description'],
+            'dependency' => $request['dependency'] ?: false,
+            'dev_mode' => $request['dev_mode'] ?: false,
+            'level_id' => $request['level_id'],
+            'order_no' => $request['order_no'],
+            'created_at' => date('Y-m-d H:i:s'),
+            'modified_at' => date('Y-m-d H:i:s')
+        ]);
+        $level_id = $request->input('level_id');
+        return redirect('/units?level_id=' . $level_id)->with(array('message' => 'Updated successfully'));
+    }
+
     public function destroy(Request $request, $id)
     {
+        $this->checkAccess(auth()->user()->isSuperAdmin() || auth()->user()->isAdmin());
         $level_id = $request->input('level_id');
         Unit::where('id', $id)->delete();
-        return redirect('/unit_views?level_id='. $level_id)->with(array('message'=> 'Deleted successfully'));
+        return redirect('/units?level_id=' . $level_id)->with(array('message' => 'Deleted successfully'));
     }
 }
