@@ -2,9 +2,15 @@
 
 namespace App\Providers;
 
-use Illuminate\Support\ServiceProvider;
+use App\ClassOfStudents;
+use App\Observers\ClassObserver;
+use Carbon\Carbon;
 use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\ServiceProvider;
+
+use Monolog\Logger;
+use Monolog\Handler\StreamHandler;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -15,13 +21,15 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        //degug mode
-        /*DB::listen(function ($query) {
-            // $query->sql
-            // $query->bindings
-            // $query->time
-            Log::info("Executed:".$query->sql);
-        });*/
+        Validator::extend('recaptcha', 'App\\Validators\\ReCaptcha@validate');
+        ClassOfStudents::observe(ClassObserver::class);
+        /** DB log */
+        /* DB::listen(function ($query) {
+            $log = ['QUERY' => $query->sql, 'TIME' => $query->time];
+            $dbLog = new Logger('DB');
+            $dbLog->pushHandler(new StreamHandler(storage_path('logs/DB-'.Carbon::now()->toDateString().'.log')));
+            $dbLog->info('DBLog', $log);
+        }); */
     }
 
     /**
