@@ -3,24 +3,41 @@
 namespace App;
 
 use Illuminate\Foundation\Auth\User as Authenticatable;
+use Illuminate\Notifications\Notifiable;
 
 class User extends Authenticatable
 {
-    /**
-     * The attributes that are mass assignable.
-     *
-     * @var array
-     */
+    use Notifiable;
+
     protected $fillable = [
-        'name', 'email', 'password', 'is_admin'
+        'name', 'email', 'password', 'role'
     ];
 
-    /**
-     * The attributes that should be hidden for arrays.
-     *
-     * @var array
-     */
     protected $hidden = [
         'password', 'remember_token',
     ];
+
+    public function isQuestionsEditor() {
+        return $this->role == 'questions_editor';
+    }
+
+    public function isAdmin() {
+        return $this->role == 'admin' || $this->role == 'superadmin';
+    }
+
+    public function isSuperAdmin() {
+        return $this->role == 'superadmin';
+    }
+
+    public function roleName() {
+        switch ($this->role) {
+            default:
+            case 'questions_editor':
+                return 'Questions Editor';
+            case 'admin':
+                return 'Admin';
+            case 'superadmin':
+                return 'Super Admin';
+        }
+    }
 }
