@@ -1,13 +1,4 @@
-/**
- * @license Copyright (c) 2003-2017, CKSource - Frederico Knabben. All rights reserved.
- * For licensing, see LICENSE.md or http://ckeditor.com/license
- */
-
 CKEDITOR.editorConfig = function( config ) {
-	// Define changes to default configuration here.
-	// For complete reference see:
-	// http://docs.ckeditor.com/#!/api/CKEDITOR.config
-
 	// The toolbar groups arrangement, optimized for a single toolbar row.
 	config.toolbarGroups = [
 		{ name: 'document',	   groups: [ 'mode', 'document', 'doctools' ] },
@@ -16,7 +7,7 @@ CKEDITOR.editorConfig = function( config ) {
 		{ name: 'forms' },
 		{ name: 'basicstyles', groups: [ 'basicstyles', 'cleanup' ] },
 		{ name: 'paragraph',   groups: [ 'list', 'indent', 'blocks', 'align', 'bidi' ] },
-		{ name: 'links' },		
+		{ name: 'links' },
 		{ name: 'insert' },
 		{ name: 'styles' },
 		{ name: 'colors' },
@@ -31,15 +22,17 @@ CKEDITOR.editorConfig = function( config ) {
 
 	// Dialog windows are also simplified.
 	config.removeDialogTabs = 'link:advanced';
-	
+
 	config.mathJaxLib = 'https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.1/MathJax.js?config=TeX-MML-AM_CHTML';
 
-    config.extraPlugins = 'filebrowser,image,sourcedialog,dialog,dialogui,font,format,chart';
+    config.extraPlugins = 'filebrowser,image,sourcedialog,dialog,dialogui,font,format,chart,tooltip';
 
-    config.filebrowserUploadUrl = '/admin/question_views/uploadImage';
+    config.filebrowserUploadUrl = '/admin/questions/uploadImage';
     config.filebrowserBrowseUrl = '/admin/js/ckeditor/plugins/imagebrowser/imagebrowser.html?imgroot=/admin/uploads';
 
     // Enable all default text formats
     config.format_tags = 'p;h1;h2;h3;h4;h5;h6;pre;address;div';
+
+	config.allowedContent = true;
 
 };
